requirements.txt
streamlit
openai
pandas
langchain
langchain-community
tiktoken
faiss-cpu
openpyxl
python-dotenv